collocation-extraction-icl/requirements.txt
pandas
pyyaml
yandex-cloud-ml-sdk
